Fixed API bug - setFontSize used int instead of double
git-svn-id: svn://svn.rforge.net/org/trunk@1105 955e8658-2bef-402f-a597-8b5c18d8b76d

+++ rosuda/pograss/PoGraSSgraphics.java
@@ -147,7 +147,7 @@ public class PoGraSSgraphics extends PoGraSS
 	lastFace=name;
 	g.setFont(currentFont=new Font(lastFace,lastFontAttr,(int)(0.5+lastFontSize)));
     };
-    public void setFontSize(int pt) {
+    public void setFontSize(double pt) {
         lastFontSize=pt;
 	g.setFont(currentFont=new Font(lastFace,lastFontAttr,(int)(0.5+lastFontSize)));
     }
